frontend/src/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.5
pydantic==2.5.3
